install script: day 3 (probably finished)
- script checks for user running the script, and SIGINT as well
- update.sh deprecated as all dotfiles will be symlinked
- added fallback function for dotfiles if "Comic Code Medium" not present
- addition: default tty font config

+++ scripts/system/reload_polybar.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+if ! hash polybar 2>/dev/null; then exit 1; fi
 
 killall -q polybar # Terminate already running bar instances
 while pgrep -u $UID -x polybar >/dev/null; do sleep 1; done # Wait until the processes have been shut down
